Use loopback devices instead of device offsets to handle partitions
A loopback devices for all partitions are added via kpartx. It is important to remove all created loopback devices before exiting the container. Otherwise, a new set of devices is added for each run until no further device node is available.

On macOS 10.13 and docker 17.12, it seems to be necessary to restart docker to get rid of the leftover devices. On Linux, it might be possible from host OS, because host and container use the same kernel. But this is untested. A reboot will likely solve it.